# backend/swagger.yaml
openapi: 3.0.3
info:
  title: AbiturAI API
  description: AI-powered DTM test prep platform for Uzbek students
  version: 3.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        firstname: { type: string }
        lastname: { type: string }
        email: { type: string }
        gender: { type: string, enum: [male, female] }
        role: { type: string, enum: [student, teacher, admin] }
        field_id: { type: integer, nullable: true }
        subject_id: { type: integer, nullable: true }
        field:
          $ref: '#/components/schemas/FieldWithSubjects'
        subject:
          $ref: '#/components/schemas/SubjectShort'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AuthResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, example: 3600 }
...
    FieldWithSubjects:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        subjects:
          type: array
          items: { $ref: '#/components/schemas/SubjectShort' }
...
    AdminDashboard:
      type: object
      properties:
        total_users: { type: integer }
        total_subjects: { type: integer }
        total_topics: { type: integer }
        total_questions: { type: integer }
        total_quiz_attempts: { type: integer }
    ValidationError:
      type: object
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
paths:
...
  /fields/{id}:
    get:
      tags: [Public]
      summary: Get single field with its subjects
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Field with subjects
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldWithSubjects' }

  # ════════════════════════════════════════
  # AUTH
  # ════════════════════════════════════════
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new student or teacher
      description: |
        **Student**: requires field_id
        **Teacher**: requires subject_id
        Admin cannot be registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstname, lastname, email, password, password_confirmation, gender, role]
              properties:
                firstname: { type: string, example: "Ali" }
                lastname: { type: string, example: "Valiyev" }
                email: { type: string, example: "ali@example.com" }
                password: { type: string, minLength: 6, example: "password123" }
                password_confirmation: { type: string, example: "password123" }
                gender: { type: string, enum: [male, female], example: "male" }
                role: { type: string, enum: [student, teacher], example: "student" }
                field_id: { type: integer, description: "Required for student", example: 2 }
                subject_id: { type: integer, description: "Required for teacher", example: 1 }
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '422':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login (email + password)
      description: Returns JWT token + user with role info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: "student@abiturai.uz" }
                password: { type: string, example: "student123" }
            examples:
              student:
                summary: Login as student
                value: { email: "student@abiturai.uz", password: "student123" }
              teacher:
                summary: Login as teacher
                value: { email: "teacher@abiturai.uz", password: "teacher123" }
              admin:
                summary: Login as admin
                value: { email: "admin@abiturai.uz", password: "admin123" }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '422':
          description: Invalid credentials
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (invalidate token)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Logged out
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh JWT token
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: New token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
  /user:
    get:
      tags: [Auth]
      summary: Get current user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }

  # ════════════════════════════════════════
  # PROFILE
  # ════════════════════════════════════════
  /profile:
    get:
      tags: [Profile]
      summary: Get current user profile with stats
      description: "Role: all authenticated. Returns user info + quiz stats + streak."
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Profile with stats
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileResponse' }
    put:
      tags: [Profile]
      summary: Update profile
      description: "Role: all authenticated. Update firstname, lastname, gender, field_id."
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname: { type: string, example: "Alisher" }
                lastname: { type: string, example: "Valiyev" }
                gender: { type: string, enum: [male, female] }
                field_id: { type: integer, nullable: true }
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: '#/components/schemas/User' }

  # ════════════════════════════════════════
  # DASHBOARD
  # ════════════════════════════════════════
  /dashboard/student:
    get:
      tags: [Dashboard]
      summary: Student dashboard with full stats
      description: "Role: student, teacher, admin. Returns quiz stats, streak, subject progress, recommendations, calendar."
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Full student dashboard
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentDashboard' }
  /dashboard/admin:
    get:
      tags: [Dashboard]
      summary: Admin dashboard stats
      description: "Role: admin only"
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Admin overview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminDashboard' }
        '403':
          description: Forbidden

  # ════════════════════════════════════════
  # STREAK & PROGRESS
  # ════════════════════════════════════════
  /streak:
    get:
      tags: [Student]
      summary: Get current streak and 30-day calendar
      description: "Role: student, teacher, admin"
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Streak data with calendar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StreakResponse' }
  /progress/subjects:
    get:
      tags: [Student]
      summary: Per-subject progress with topic breakdown
      description: "Role: student, teacher, admin. Shows subjects for student's field with per-topic progress."
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Subject progress array
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/SubjectProgress'
                    - type: object
                      properties:
                        topics:
                          type: array
                          items:
                            type: object
                            properties:
                              topic_id: { type: integer }
                              topic_title: { type: string }
                              status: { type: string, enum: [new, in_progress, completed] }
                              best_score: { type: integer, nullable: true }
                              attempts_count: { type: integer }
  /recommendations:
    get:
      tags: [Student]
      summary: AI-powered weak topic recommendations
      description: "Role: student, teacher, admin. Returns top 5 topics to focus on (unattempted or lowest score)."
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Recommended topics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Recommendation' }

  # ════════════════════════════════════════
  # SUBJECTS
  # ════════════════════════════════════════
  /subjects:
    get:
      tags: [Subjects]
      summary: List all subjects with topic count
      description: "Role: all authenticated"
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Subject' }
    post:
      tags: [Subjects]
      summary: Create subject
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "Informatika" }
                description: { type: string, example: "DTM informatika fani" }
      responses:
        '201':
          description: Subject created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subject' }
  /subjects/{id}:
    put:
      tags: [Subjects]
      summary: Update subject
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '200':
          description: Subject updated
    delete:
      tags: [Subjects]
      summary: Delete subject
      description: "Role: admin only"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Subject deleted
  /subjects/{id}/topics:
    get:
      tags: [Subjects]
      summary: List topics in a subject
      description: "Role: all authenticated"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Topics list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TopicShort' }

  # ════════════════════════════════════════
  # TOPICS
  # ════════════════════════════════════════
  /topics:
    post:
      tags: [Topics]
      summary: Create topic
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject_id, title, content]
              properties:
                subject_id: { type: integer, example: 1 }
                title: { type: string, example: "Integrallar" }
                content: { type: string }
                order_num: { type: integer, example: 4 }
      responses:
        '201':
          description: Topic created
  /topics/{id}:
    get:
      tags: [Topics]
      summary: Get topic content (lesson page)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Topic with content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Topic' }
    put:
      tags: [Topics]
      summary: Update topic
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject_id: { type: integer }
                title: { type: string }
                content: { type: string }
                order_num: { type: integer }
      responses:
        '200':
          description: Topic updated
    delete:
      tags: [Topics]
      summary: Delete topic
      description: "Role: admin only"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Topic deleted

  # ════════════════════════════════════════
  # QUESTIONS
  # ════════════════════════════════════════
  /questions:
    get:
      tags: [Questions]
      summary: List questions (paginated, filterable)
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: topic_id
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200':
          description: Paginated questions
    post:
      tags: [Questions]
      summary: Create question
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic_id, question_text, option_a, option_b, option_c, option_d, correct_answer]
              properties:
                topic_id: { type: integer, example: 1 }
                question_text: { type: string, example: "log_2(16) = ?" }
                option_a: { type: string, example: "2" }
                option_b: { type: string, example: "3" }
                option_c: { type: string, example: "4" }
                option_d: { type: string, example: "8" }
                correct_answer: { type: string, enum: [A, B, C, D], example: "C" }
                explanation: { type: string }
                difficulty: { type: string, enum: [easy, medium, hard], default: medium }
      responses:
        '201':
          description: Question created
  /questions/{id}:
    get:
      tags: [Questions]
      summary: Get single question
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Question details
    put:
      tags: [Questions]
      summary: Update question
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question_text: { type: string }
                option_a: { type: string }
                option_b: { type: string }
                option_c: { type: string }
                option_d: { type: string }
                correct_answer: { type: string, enum: [A, B, C, D] }
                explanation: { type: string }
                difficulty: { type: string, enum: [easy, medium, hard] }
      responses:
        '200':
          description: Question updated
    delete:
      tags: [Questions]
      summary: Delete question
      description: "Role: admin only"
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Question deleted

  # ════════════════════════════════════════
  # QUIZ
  # ════════════════════════════════════════
  /quiz/{topicId}/start:
    post:
      tags: [Quiz]
      summary: Start quiz (10 random questions)
      description: "Role: student, teacher, admin. Returns questions WITHOUT correct answers."
      security: [{ bearerAuth: [] }]
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Quiz questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic:
                    type: object
                    properties:
                      id: { type: integer }
                      title: { type: string }
                  questions:
                    type: array
                    items: { $ref: '#/components/schemas/QuestionForQuiz' }
  /quiz/submit:
    post:
      tags: [Quiz]
      summary: Submit quiz + get AI diagnosis
      description: "Also records streak + topic progress."
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic_id, answers]
              properties:
                topic_id: { type: integer, example: 1 }
                answers:
                  type: array
                  items:
                    type: object
                    required: [question_id, selected_answer]
                    properties:
                      question_id: { type: integer }
                      selected_answer: { type: string, enum: [A, B, C, D] }
      responses:
        '200':
          description: Quiz result
          content:
            application/json:
              schema:
                type: object
                properties:
                  attempt_id: { type: integer }
                  score: { type: integer }
                  total: { type: integer }
                  percentage: { type: number }
                  diagnosis:
                    type: array
                    nullable: true
                    items: { $ref: '#/components/schemas/DiagnosisItem' }
  /quiz/result/{id}:
    get:
      tags: [Quiz]
      summary: Get detailed quiz result
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Full quiz attempt with answers
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizAttempt' }
  /quiz/history:
    get:
      tags: [Quiz]
      summary: Get quiz history (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
      responses:
        '200':
          description: Paginated quiz attempts

  # ════════════════════════════════════════
  # AI FEATURES (Stateless)
  # ════════════════════════════════════════
  /tutor/ask:
    post:
      tags: [AI Features]
      summary: Ask AI tutor (stateless, no chat history)
      description: "RAG-powered. For persistent chat, use /tutor/chat endpoints."
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string, maxLength: 1000, example: "Logarifm xossalarini tushuntiring" }
      responses:
        '200':
          description: AI answer with sources
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TutorResponse' }
  /feynman/evaluate:
    post:
      tags: [AI Features]
      summary: Evaluate student explanation (Feynman method)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic_id, explanation]
              properties:
                topic_id: { type: integer, example: 1 }
                explanation: { type: string, minLength: 20 }
      responses:
        '200':
          description: Evaluation result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeynmanResult' }
  /generator/create:
    post:
      tags: [AI Features]
      summary: Generate DTM questions from text
      description: "Role: teacher, admin"
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 50 }
                count: { type: integer, minimum: 1, maximum: 10, default: 5 }
                topic_id: { type: integer, nullable: true }
      responses:
        '200':
          description: Generated questions (preview)
        '201':
          description: Generated and saved

  # ════════════════════════════════════════
  # CHAT (Session-based AI Tutor)
  # ════════════════════════════════════════
  /tutor/history:
    get:
      tags: [Chat]
      summary: List chat sessions
      description: "Role: student, teacher, admin. Returns last 20 sessions with message count."
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Chat sessions list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChatSession' }
  /tutor/chat:
    post:
      tags: [Chat]
      summary: Create new chat session
      description: "Optionally pass subject_id for context filtering."
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject_id: { type: integer, nullable: true, example: 1 }
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatSession' }
  /tutor/chat/{id}:
    get:
      tags: [Chat]
      summary: Get chat session with all messages
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Session with messages
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ChatSession'
                  - type: object
                    properties:
                      messages:
                        type: array
                        items: { $ref: '#/components/schemas/ChatMessage' }
    delete:
      tags: [Chat]
      summary: Delete chat session
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Session deleted
  /tutor/chat/{id}/ask:
    post:
      tags: [Chat]
      summary: Send message in chat session
      description: "Saves user message, calls RAG + Gemini with conversation context, returns both messages."
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: User message + AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatAskResponse' }

  # ════════════════════════════════════════
  # TEACHER
  # ════════════════════════════════════════
  /teacher/dashboard:
    get:
      tags: [Teacher]
      summary: Teacher dashboard stats
      description: "Role: teacher, admin. Stats for teacher's assigned subject."
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Teacher dashboard
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeacherDashboard' }
        '400':
          description: No subject assigned
...